backend/photo-processor/requirements.txt
functions-framework==3.5.0
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.18.0
google-api-python-client==2.114.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
Pillow==10.2.0
pillow-heif==0.15.0

